feat(qcow2): Q2R3 recipes regenerate compressed clusters instead of storing them
A QCOW2 image decomposes into a recipe (every non-member byte) plus the
rankimg member stream. Q2R2 recorded the compressed-cluster byte length in
the recipe table but still copied the whole deflate stream into the recipe
gap, so every compressed cluster was stored TWICE: once in rankimg (as raw
bytes) and once in the recipe (as the original stream).

Q2R3 records the deflate parameters instead of the bytes. At strip time the
pack inflates each cluster, asks deflate_repro for a bit-exact
(level, memLevel, strategy, engine), and -- only when the re-encode
reproduces the original stream byte-for-byte -- leaves it out of the
recipe. rebuild and the read path re-encode it on demand. Anything the
finder cannot reproduce stays verbatim, so the output is bit-exact either
way. On ubuntu2404-base.qcow2 all 28029 compressed clusters qualify: the
recipe drops 625 MB while the image still reads back identical.

The map format grows a v2 wire ("MRM2", 40-byte entries) so a stored
cluster with no bytes anywhere can be named: kind 2 (REPRO) points at the
raw member range plus the deflate parameters, and cpack_map_serve
regenerates the stream. v1 containers keep the exact old bytes on disk.

Migration is pack-agnostic and needs no flag: a map records the pack
generation that produced it (decomp_gen, FS-owned header field), and when
a pack bumps its generation the sweep re-derives that decomposition from
the RECONSTRUCTED container stream -- vol_read_file serves the original
bytes through the old map, and the pack strips them again with its current
recipe format. Idempotent, and any pack (tar/zip/p7z/rawdisk) migrates
through the same code when its generation moves.

Fixes found while measuring this on the real 10 GB volume:

- vol_cpack_read returned the served byte count through an `int`, so any
  container over 2 GiB read back as a negative length and every ranged read
  of a large member failed. Now int64_t.
- cpack_map_get re-loaded a member's recipe (a B+ tree walk plus a hash of
  the whole blob) on every member read. A 28k-cluster guard pass turned
  into 71 GB of reads and an apparent hang. Recipes are now cached in two
  slots keyed by address -- the address IS the BLAKE3 of the bytes, so a
  hit can never be stale and there is nothing to invalidate.
- blkio_looks_like_device() treated every "/dev/" path as a block device,
  so no volume image under /dev/shm could be opened by path at all
  (BLKGETSIZE64 on tmpfs returns ENOTTY). It now stats and asks for
  S_ISBLK.
- A container commit re-creates its members under the same names, so a
  record replaced mid-sweep left the walk holding a retired inode id and
  every sibling reported a hard failure. The v3 sweep path re-resolves the
  name before dispatching.
- Migrating a container whose member is itself a decomposed container left
  the nested branch behind: the fresh record inherited the CONTAINER stamp
  while the old map's members stayed on disk. Members' own siblings are now
  purged before the rewrite.

Verified on ubuntu2404-base.qcow2 (625 MB, 28k compressed clusters):
invf-verify --deep reads 11/11 files, 10021843507 bytes, 0 corrupt; the
container reads back byte-identical to the source; the map is MRM2
gen=2 with 28029 REPRO entries. make test and tools/test-qcow2.sh pass.

Also adds the invf-sweep TUI stage/dedupe-scope counters (cross/intra/
merged) that were already in flight, with tools/test-sweep-ui.sh.

+++ src/core/vol_read.c
@@ -1643,13 +1643,13 @@ int vol_read_range(invfs_volume *v, uint64_t inode_id, uint64_t offset,
                 char mbn[288];
                 snprintf(mbn, sizeof mbn, "%s!mbrmap", rname);
                 if (vol_find(v, mbn) != 0) {
-                    int mrc;
+                    int64_t mrc;
                     if (offset >= ast_h.file_size) { free(rec); return 0; }
                     mrc = cpack_map_read(v, rname, inode_id,
                                          (uint64_t)ast_h.file_size, offset,
                                          (uint8_t *)buf, len);
                     free(rec);
-                    return mrc;
+                    return (int)mrc;
                 }
                 whole = seek;   /* map missing: the exec rebuild answers */
             }
